requirements.txt
torch>=2.0
torchvision
opencv-python
numpy
matplotlib
Pillow
pywavelets
scikit-image
tqdm
pandas
